refactor: 키를 가져오는 역할 위임
