+++ src/main/java/koreatech/in/repository/AuthenticationMapper.java
@@ -12,5 +12,5 @@ public interface AuthenticationMapper {
     void deleteRefreshToken(Integer userId);
     //TODO 사장님 인증 + 메모리DB 부분도 여기로 추출할 것
 
-    String getKey() throws IOException;
+    String getDeprecatedKey(String keyName) throws IOException;
 }
